add csr query
